% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitTomato.R
\docType{package}
\name{bitTomato-package}
\alias{bitTomato}
\alias{bitTomato-package}
\title{Text data of articles published in News Tomato}
\description{
Text data created from news articles provided by the media
company NewsTomato. It includes examples of text analysis using this data,
as well as functions and vignettes.
}
\details{
\tabular{ll}{
Package: \tab bitTomato\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2023-07-22\cr
License: \tab GPL-3\cr
}
}
\author{
유충현
Maintainer: 유충현 \href{mailto:choonghyun.ryu@gmail.com}{choonghyun.ryu@gmail.com}
}
