% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherry_tomato.R
\docType{data}
\name{cherry_tomato}
\alias{cherry_tomato}
\title{방울토마토 - 뉴스토마토 뉴스 기사 샘플 데이터}
\format{
300 관측치와 8개의 변수를 포함하는 tibble 객체임. 변수들은 다음과 같음.:
\describe{
\item{title}{뉴스 기사 제목}
\item{url}{뉴스 기사 원본이 게시된 URL 주소}
\item{contents_origin}{정제되지 않은 뉴스 기사}
\item{contents}{어느 정도 정제된 뉴스 기사}
\item{author}{뉴스 기사를 작성한 저자나 기관}
\item{create_dt}{뉴스 기사를 생성한 일시.}
\item{category}{뉴스 기사의 카테고리}
\item{subcategory}{뉴스 기사의 서브 카테고리.}
}
}
\source{
{
"뉴스토마토 뉴스 기사 샘플링" in \url{https://www.newstomato.com/}, License : GPL-3
}
}
\usage{
data(cherry_tomato)
}
\description{
간단한 분석 실습을 위한 뉴스토마토의 뉴스의 샘플 데이터
}
\details{
기자가 "인포머셜" 제외하고 정치, 산업, 사회 3개 카테고리 기준으로 뉴스 본문의
길이가 25분위에서 75분위 사이 각각 100개 뉴스기사를 추출함.
데이터셋의 이름은 작은 크기의 데이터셋이므로 방울토마토(cherry tomato)로 명명함.
}
\keyword{datasets}
