% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tomato.R
\docType{data}
\name{tomato}
\alias{tomato}
\title{뉴스토마토 뉴스 기사 데이터}
\format{
164267 관측치와 각각 7개의 변수를 포함하는 tibble 객체임. 변수들은 다음과 같음.:
\describe{
\item{title}{뉴스 기사 제목}
\item{url}{뉴스 기사 원본이 게시된 URL 주소}
\item{contents_origin}{정제되지 않은 뉴스 기사}
\item{contents}{어느 정도 정제된 뉴스 기사}
\item{author}{뉴스 기사를 작성한 저자나 기관}
\item{create_dt}{뉴스 기사를 생성한 일시.}
\item{category}{뉴스 기사의 카테고리}
\item{subcategory}{뉴스 기사의 서브 카테고리.}
}
}
\source{
{
"뉴스토마토 뉴스 기사 샘플링" in \url{https://www.newstomato.com/}, License : GPL-3
}
}
\usage{
data(tomato)

data(tomato_origin)
}
\description{
언론사인 뉴스토마토의 뉴스 기사로 정제되지 않은 기사와 정제된 기사를 포함
}
\details{
tomato_origin 데이터셋의 기사는 정제되지 않은 기사로 HTML 태그 등이 포함되어 있으며,
tomato 데이터셋의 기사는 어느 정도 정제된 기사로 HTML 태그 등이 어느 정도 정제되어 있음
}
\keyword{datasets}
